Algorithm skeleton for #36
VariableDefinitions created:

- include top tagging
- subjettiness accessors
- kt splitting scale accessors
- working point definitions

We want to do our best in retrieving the substructure variables. If the decorations don't exist, we will compute them and add them. If we do not have access to constituents, that's a problem with the data we're looking at.

Also switch to using getAttribute() for jets and SG::AuxElement::Decorator for setting.